+++ packages/formfield/mwc-formfield-base.ts
@@ -109,7 +109,7 @@ export class FormfieldBase extends BaseElement {
       </div>`;
   }
 
-  private _labelClick() {
+  protected _labelClick() {
     const input = this.input;
     if (input) {
       input.focus();